Speed up the commons functions
My 10M (sqlite3_opt_batched) run goes from 42s to 26s.

Switching to https://docs.python.org/3/library/os.html#os.urandom would
make this even faster, but I don't think that's the point of the
exercise so leaving it at this. I think it's still nice to have because
it reduces the time by a lot and makes it _faster for me to benchmark_